// src/api/rest/v1/gen/models/FeedbackCreate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Gamanza REST recommendation engine API specification
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: info@gamanza.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * 
 * @export
 * @interface FeedbackCreate
 */
export interface FeedbackCreate {
    /**
     * 
     * @type {number}
     * @memberof FeedbackCreate
     */
    gameId: number;
    /**
     * 
     * @type {string}
     * @memberof FeedbackCreate
     */
    gameName: string;
    /**
     * 
     * @type {string}
     * @memberof FeedbackCreate
     */
    gameCategory: FeedbackCreateGameCategoryEnum;
    /**
     * 
     * @type {string}
     * @memberof FeedbackCreate
     */
    playerId: string;
    /**
     * 
     * @type {string}
     * @memberof FeedbackCreate
     */
    feedbackType: FeedbackCreateFeedbackTypeEnum;
}


/**
 * @export
 */
export const FeedbackCreateGameCategoryEnum = {
    Slots: 'SLOTS',
    DiceGames: 'DICE_GAMES',
    Blackjack: 'BLACKJACK',
    Roulette: 'ROULETTE',
    Baccarat: 'BACCARAT',
    Poker: 'POKER',
    Other: 'OTHER'
} as const;
export type FeedbackCreateGameCategoryEnum = typeof FeedbackCreateGameCategoryEnum[keyof typeof FeedbackCreateGameCategoryEnum];

/**
 * @export
 */
export const FeedbackCreateFeedbackTypeEnum = {
    GameStarted: 'GameStarted',
    GameRated: 'GameRated',
    GamePlayed: 'GamePlayed',
    GameLiked: 'GameLiked'
} as const;
export type FeedbackCreateFeedbackTypeEnum = typeof FeedbackCreateFeedbackTypeEnum[keyof typeof FeedbackCreateFeedbackTypeEnum];